fix(db): global Sengoku/Neo Coby/Neo Helmep wrong ID
We have found that we have screwed-up their id, Sengoku should have been 5054 instead of 5058, Neo Coby 5054-5055 instead of 5053-5054 and Neo Helmep 5056 instead of 5055.

So we have this correction and moved them to their right IDs. However this modification is a breaking change as we haven't find a way to make it smooth, sorry for the inconvenience. So, if you have add them on your box, your must delete them and re-add them manually.