feat(sugocleaner): rework lists
Since game update 11.2.2, we can now increased stats even when they are locked. So some of the SugoCleaner list have lost their purpose. In order to keep the app simple to use, we have replaced them with a single "Waiting list", which allow you to put away the character you don't have decided what to do with them.